+++ src/scooze/config.py
@@ -1,7 +1,19 @@
 import importlib.metadata
+import os
+from pathlib import Path
 from typing import NamedTuple
 
 from pydantic_settings import BaseSettings
+
+# region Flags
+
+DEBUG = os.getenv("SCOOZE_DEBUG", "False") == "True"
+PACKAGE_ROOT = Path(__file__).parent
+DEFAULT_BULK_FILE_DIR = Path.home() / ".scooze" / "data" / "bulk"
+DEFAULT_DECKS_DIR = Path.home() / ".scooze" / "data" / "decks"
+DEFAULT_LOGS_DIR = Path.home() / ".scooze" / "logs"
+
+# endregion
 
 
 class Version(NamedTuple):
@@ -15,6 +27,11 @@ class ScoozeSettings(BaseSettings):
     mongo_dsn: str = "mongodb://127.0.0.1:27017"
     mongo_db: str = "scooze"
 
+    debug: bool = DEBUG
+    package_root: Path = PACKAGE_ROOT
+    bulk_file_dir: Path = DEFAULT_BULK_FILE_DIR
+    decks_dir: Path = DEFAULT_DECKS_DIR
+    logs_dir: Path = DEFAULT_LOGS_DIR
     testing: bool = False
 
     @property