+++ src/scooze/console/commands/load/cards.py
@@ -5,7 +5,7 @@ from cleo.helpers import option
 from scooze.api import ScoozeApi
 from scooze.bulkdata import download_bulk_data_file_by_type
 from scooze.catalogs import ScryfallBulkFile
-from scooze.utils import DEFAULT_BULK_FILE_DIR
+from scooze.config import CONFIG
 
 
 class LoadCardsCommand(Command):
@@ -21,7 +21,7 @@ class LoadCardsCommand(Command):
         option(
             "bulk-data-dir",
             description="Directory to load bulk files from.",
-            default=DEFAULT_BULK_FILE_DIR,
+            default=CONFIG.bulk_file_dir,
             value_required=True,
             flag=False,
         ),