+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 - Refactored the bulkdata API endpoints ([#279](https://github.com/arcavios/scooze/pull/279))
 - Renamed DeckPart to CardList ([#281](https://github.com/arcavios/scooze/pull/281))
 - Use `pathlib` for consistent cross-OS file paths ([#222](https://github.com/arcavios/scooze/pull/222))
+- Default to home/scooze for storage of bulk data and logs ([#294](https://github.com/arcavios/scooze/pull/294))
+
 
 #### Fixed